+++ src/main/java/dev/nullpanic/tamagotchitelegrambot/command/CommandContainer.java
@@ -1,9 +1,7 @@
 package dev.nullpanic.tamagotchitelegrambot.command;
 
 import com.google.common.collect.ImmutableMap;
-import dev.nullpanic.tamagotchitelegrambot.command.petcommand.CreatePetCommand;
-import dev.nullpanic.tamagotchitelegrambot.command.petcommand.FeedPetCommand;
-import dev.nullpanic.tamagotchitelegrambot.command.petcommand.GetAllPetsCommand;
+import dev.nullpanic.tamagotchitelegrambot.command.petcommand.*;
 import dev.nullpanic.tamagotchitelegrambot.service.PetService;
 import dev.nullpanic.tamagotchitelegrambot.service.SendBotMessageService;
 import dev.nullpanic.tamagotchitelegrambot.service.TelegramUpdateService;
@@ -23,12 +21,13 @@ public class CommandContainer {
 
         commandMap = ImmutableMap.<String, Command>builder()
                 .put(CommandName.START.getCommandName(), new StartCommand(sendBotMessageService, telegramUserService))
-                .put(CommandName.STOP.getCommandName(), new StopCommand(sendBotMessageService, telegramUserService))
+                .put(CommandName.STOP_ALL.getCommandName(), new StopAllPetsCommand(sendBotMessageService, telegramUpdateService, petService))
                 .put(CommandName.HELP.getCommandName(), new HelpCommand(sendBotMessageService))
                 .put(CommandName.STAT.getCommandName(), new StatCommand(sendBotMessageService, telegramUserService))
                 .put(CommandName.CREATE.getCommandName(), new CreatePetCommand(sendBotMessageService, petService, telegramUpdateService))
                 .put(CommandName.FEED.getCommandName(), new FeedPetCommand(sendBotMessageService, petService, telegramUpdateService))
                 .put(CommandName.GET_ALL_PETS.getCommandName(), new GetAllPetsCommand(sendBotMessageService, petService, telegramUpdateService))
+                .put(CommandName.STOP.getCommandName(), new StopPetCommand(sendBotMessageService, telegramUpdateService, petService))
                 .build();
 
         unknownCommand = new UnknownCommand(sendBotMessageService);